set state pda as a client property
